Clean up target folders from source directory
Scripts such as the importer do still need access to a JAR file
with certain executable Java classes produced during compilation.
By default the scripts search for this file in the place where Maven
leaves it during compilation, so this is the most convenient place
to install it after cleaning up intermediate files in this directory.