Maintain jar file needed by scripts in PORTAL_HOME
Since the compiled Java classes needed by scripts such as the
importer are no longer copied in the war file deployed as a web
application, a second compiled file now needs to be maintained in
order to use these scripts. By default the scripts search for the
jar file in the place where Maven puts it during compilation, so
this is the most convenient place to install it.