Re-enable MSKCC linkouts, fixing patch failure
Something had changed in the vicinity of the links disabled by the
portal config patch, making it fail to apply to versions of the
example file in development branches. Since there was no good reason
to disable these linkouts in the first place, the incompatibility
can be solved by removing the hunk from the patch.