Adjust testing commands to the new name of the war
The .war file produced when compiling cBioPortal has had a version
number in the name for a while now, but the commands suggested for
developers who want to build and test a local or git-hosted branch had
yet to be adjusted.

In addition, unzipping the file instead of letting Tomcat manage
it will make it easier to mount and edit deployed files without
recompiling.