Remove opening a DB port from the basic workflow
Opening up a port on the database container is only relevant to
workflows where users want to connect external tools to the database
server. Keeping it closed unless there's a specific reason to makes
the steps simpler to understand and avoids potential security issues.

The option is still documented, as well as a method to troubleshoot
the database if the server was started without ports open to the
outside.