Harmonise outline of README Markdown headings
The README file skipped over the second and third level of headings in
some places, and a section that is likely going to be needed right after
installation was only mentioned in a very short paragraph all the way at
the bottom.  Reorganising this outline should make it more comfortable
to follow along the instructions.