Link to the master version of the seed DB README
The README file for the seed database files on Datahub used to apply to
only one release of cBioPortal, but it now documents clearly which
cBioPortal releases are supported by it. This means that linking to a
newer version of this README file will no longer cause the setup to
fail, and there is no longer a need to specifically suggest a very old
(upwards-migratable) seed database.