Make Dockerfile.local build skip .git folder
In recent versions of cBioPortal, the code to provision the `info` API
call looks at the Git metadata folder while building the portal. This
folder is not included in the development images, as it generally
contains the full history of many branches not relevant to this
particular build. Disabling the feature should allow the build to
continue.