Remove warning not to use an existing db
This was vital and led to confusion when we documented using the
MySQL entrypoint scripts to first load the seed database. Now that
we made that an explicit separate step, I don't think this warning
serves a purpose other than to confuse people who have previously
created a database folder and want to connect a fresh containerised
MySQL server to it.